Fix type of _cairo_memory_stream_destroy parameter
In 9b9952ab4f64a476bbadfd3e6cf1e66c1137ccd7
_cairo_memory_stream_destroy was changed to take an unsigned long
instead of unsigned int, and the two callsites in cairo-gl-shaders.c
weren't updated.

+++ src/cairo-gl-shaders.c
@@ -632,7 +632,7 @@ cairo_gl_shader_get_vertex_source (cairo_gl_var_type_t src,
 {
     cairo_output_stream_t *stream = _cairo_memory_stream_create ();
     unsigned char *source;
-    unsigned int length;
+    unsigned long length;
     cairo_status_t status;
 
     cairo_gl_shader_emit_variable (stream, src, CAIRO_GL_TEX_SOURCE);
@@ -768,7 +768,7 @@ cairo_gl_shader_get_fragment_source (GLuint tex_target,
 {
     cairo_output_stream_t *stream = _cairo_memory_stream_create ();
     unsigned char *source;
-    unsigned int length;
+    unsigned long length;
     cairo_status_t status;
 
     cairo_gl_shader_emit_color (stream, tex_target, src, CAIRO_GL_TEX_SOURCE);
